basic part of task 1 that impletement of system call has been finished and successfully output the result! it remains to finish the error handling and returning.

+++ user/test_getprocs.c
@@ -1,8 +1,7 @@
 #include "user/user.h"
-#include "kernel/types.h"
 
 int main(int argc, char *argv[]){
-    int count = getprocs();
+    unsigned long count = getprocs();
     printf("There are %d active processes.\n",count);
     exit(0);
 }
